Readding bullets

+++ Orbeeto/Components/PlayerGun.hpp
@@ -4,6 +4,6 @@
 
 struct PlayerGun : Component {
 	uint32_t* owner;  // Pointer to player entity
-	float cooldown;  // The base rate at which the gun fires bullets
+	int cooldown;  // The base rate at which the gun fires bullets
 	float heatDissipation;
 };
